# src/docs/swagger.yaml
definitions:
  models.APIError:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  models.AdminAuth:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.AdminInfo:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  models.ConnectRequest:
    properties:
      token:
        type: string
      user_id:
        type: string
    type: object
  models.ConnectResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
  models.TokenCreateResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
    type: object
  models.TokenInfo:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_uses:
        type: integer
      token_hash:
        type: string
      usage_count:
        type: integer
      user_id:
        type: string
    type: object
  models.UserCreateRequest:
    properties:
      email:
        type: string
      inputs_format:
        type: string
      model_type:
        type: string
      outputs_format:
        type: string
      username:
        type: string
    type: object
  models.UserCreateResponse:
    properties:
      id:
        type: string
    type: object
  models.UserInfo:
    properties:
      created_at:
        type: string
      email:
        type: string
      inputs_format:
        type: string
      model_type:
        type: string
      outputs_format:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /admins/:
    get:
      consumes:
      - application/json
      description: get admins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AdminInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: get admins
      tags:
      - admins
  /admins/invite:
    post:
      consumes:
      - application/json
      description: invite admin
      parameters:
      - description: Admin Email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: invite admin
      tags:
      - admins
  /admins/login:
    post:
      consumes:
      - application/json
      description: login admin
      parameters:
      - description: Admin Auth
        in: body
        name: AdminAuth
        required: true
        schema:
          $ref: '#/definitions/models.AdminAuth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: login admin
      tags:
      - admins
  /admins/signup:
    post:
      consumes:
      - application/json
      description: sign up admin
      parameters:
      - description: Admin Auth
        in: body
        name: AdminAuth
        required: true
        schema:
          $ref: '#/definitions/models.AdminAuth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: sign up admin
      tags:
      - admins
  /tokens/:
    get:
      consumes:
      - application/json
      description: get tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TokenInfo'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: get tokens
      tags:
      - tokens
  /tokens/create:
    post:
      consumes:
      - application/json
      description: create token
      parameters:
      - description: Client ID
        in: body
        name: client_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: create token
      tags:
      - tokens
  /users/:
    get:
      consumes:
      - application/json
      description: get users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: get users
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: get user by ID
      tags:
      - users
  /users/connect:
    post:
      consumes:
      - application/json
      description: connect
      parameters:
      - description: Connect Request
        in: body
        name: ConnectRequest
        required: true
        schema:
          $ref: '#/definitions/models.ConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConnectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: connect
      tags:
      - users
  /users/create:
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: User Create Request
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: create user
      tags:
      - users
swagger: "2.0"